refactor(checker): split assignment_checker_tests.rs under 2000 LOC (arch-guard)
Move the tuple-passthrough and callable-union assignment tests into a new
sibling module assignment_checker_tuple_callable_tests.rs so the parent file
drops from 2035 to 1893 lines. Pure test relocation; no behavior change.
